requirements.txt
pillow>=12.1.1
python-docx>=1.2.0
pyinstaller>=6.19.0
psutil>=7.2.2